revert pkg/apis/componentconfig/types.go; add pkg/kubeapiserver/apis/kubeapiserverconfig/types.go
